refactor(auth): improve jwt validation
