+++ src/api/constants/apiRoutes/plugins/pluginsRoutes.ts
