+++ src/api/models/index.ts
@@ -1,6 +1,3 @@
-export * from './apiRoutes/auth/authRoutes';
-export * from './apiRoutes';
-export * from './apiRoutes/profile/profileroutes';
 export * from './mongoose';
 export * from './mongoose/plugins/plugins';
 export { default as referred } from './mongoose/referred/referred';
