+++ src/api/models/mongoose/championship/championship.ts
