+++ src/api/models/mongoose/championship/configuration.ts
