+++ src/api/models/mongoose/championship/group.ts
