+++ src/api/models/mongoose/championship/invitationLink.ts
