+++ src/api/models/mongoose/championship/match.ts
