+++ src/api/models/mongoose/championship/registration.ts
