+++ src/api/models/mongoose/championship/statistics.ts
