+++ src/api/validators/championships/championship.validator.ts
